bug fixes
- hasn't been used in interactive mode in 8-9 months. fix some things
  that have cropped up

+++ scripts/win-host-setup.ps1
@@ -14,11 +14,16 @@ function Get-Script {
         [switch]$Run = $false
     )
 
-    $cli = New-Object System.Net.WebClient
-    $cli.DownloadString("https://raw.githubusercontent.com/PlayOneMoreGame/computer-setup/$Branch/intune/$ScriptName") | Out-File $ScriptName
+    if ((Test-Path $ScriptName) -eq $false) {
+        $dir = Split-Path -Path $ScriptName
+        mkdir $dir
+
+        $cli = New-Object System.Net.WebClient
+        $cli.DownloadString("https://raw.githubusercontent.com/PlayOneMoreGame/computer-setup/$Branch/intune/$ScriptName") | Out-File $ScriptName
+    }
 
     if ($Run -eq $true) {
-        ./$ScriptName
+        . $ScriptName
     }
 }
